Combine admin forms
Put each old admin form into the same one, split with tabs
