+++ CSIS3540Project/CSIS3540Project.csproj
@@ -48,11 +48,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AdminRooms.cs">
+    <Compile Include="AdminForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="AdminRooms.Designer.cs">
-      <DependentUpon>AdminRooms.cs</DependentUpon>
+    <Compile Include="AdminForm.Designer.cs">
+      <DependentUpon>AdminForm.cs</DependentUpon>
     </Compile>
     <Compile Include="LoginForm.cs">
       <SubType>Form</SubType>
@@ -62,12 +62,6 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="AdminStudents.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="AdminStudents.Designer.cs">
-      <DependentUpon>AdminStudents.cs</DependentUpon>
-    </Compile>
     <Compile Include="StudentsReservations.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -79,14 +73,11 @@
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="AdminRooms.resx">
-      <DependentUpon>AdminRooms.cs</DependentUpon>
+    <EmbeddedResource Include="AdminForm.resx">
+      <DependentUpon>AdminForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="LoginForm.resx">
       <DependentUpon>LoginForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="AdminStudents.resx">
-      <DependentUpon>AdminStudents.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="StudentsReservations.resx">
       <DependentUpon>StudentsReservations.cs</DependentUpon>
